\name{fasttree}
\title{Simulate and Sample From Continuous Trait on a Birth/Death Tree}
\usage{
fasttree (n, lambda, mu, traits = FALSE, sigma = 1,
  alpha = 0, treeDepth = 1, sampleTimes = 1, seed = 1L,
  traitseed = NULL, startTrait = 0)
}
\arguments{
\item{n}{Number of extant (non-extinct) taxa at the tips of the generated tree}

\item{lambda}{Speciation rate}

\item{mu}{Extinction rate}

\item{traits}{If false, the function just returns the generated birth/death tree. If true, returns a list of the tree plus trait samples (see below for the format.}

\item{sigma}{The rate at which the simulated trait varies}

\item{alpha}{For alpha=0, the trait will be evolve under a Brownian motion model (i.e. simple random walk). For alpha > 0, this is the parameter for the Ornstein-Uhlenbeck process, controlling how strongly the trait is biased to return towards zero.}

\item{treeDepth}{Once generated, the tree will be scaled so that the tip to crown depth is treeDepth.}

\item{sampleTimes}{A list (between 0 and treeDepth) of times at which to record the trait values. Small values correspond to times close to the crown, large values close to the present time.}

\item{seed}{Random number seed for generating the birth/death tree.}

\item{traitseed}{Random number seed for evolving the random trait. If NULL, the value of 'seed' will be used.}

\item{startTrait}{The value of the trait to be evolved, at the crown of the tree.}
}
\description{
...
}
\details{

The birth/death tree is generated backwards in time: starting with the
desired number of tips, a sequence of extinction and speciation events
(in ratio mu:lambda) are generated, until the number of taxa has
reduced to one (the crown of the tree.) A consequence of this is that
if we demand (e.g.) a 100 taxon tree, it may have had more than 100
taxa in the recent past. Once the sequence of speciations/extinctions
have been determined, we work forward in time to simulate the timings
of the speciation/extinction events, and which taxon they occur on.
Finally the tree is scaled to give crown (the deepest speciation
event) to tip distance given by the treeDepth trait.

Note that only the ratio between mu and lambda has effect, and that
there is no guarantee that both lineages from the crown speciation
event will have extant descencents at the tips of the tree. The random
tree generation is seeded by parameter 'seed'.

After constructing the tree, a continuous trait is simulated on the
tree. At times specified by sampleTimes, all trait values in taxa
extant at that time are returned. The trait is generated by an
Ornstein-Uhlenbeck process, with parameters sigma (controlling how
quickly the trait changes) and alpha (controlling how strongly the
trait is attracted to zero.) Setting alpha=0 gives a Brownian motion
process (a random walk unaffected by the current trait value.) Trait
simulation is seeded by parameter 'traitseed', which defaults to
'seed'. Should you wish to simulate several traits over the same tree,
supply the same 'seed' value but differing 'traitseed' values over
several calls.

The function returns a list of four values. $tree is the simulated
tree (in the ape packages 'phylo' format), $samples is a two
dimensional table of trait values, with columns indexed by
sampleTimes), $n.leaves is an integer vector of how many taxa were
extant at each sampleTime, and $sampleTimes is a copy of the
sampleTimes parameter passed to the function.
}

